requirements.txt
# Core
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# NLP & Extraction
spacy==3.7.2
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.0

# Graph Database
neo4j==5.14.0
py2neo==2021.2.3

# LLM & RAG
langchain==0.1.0
anthropic==0.8.0
openai==1.6.0

# Vector Store
chromadb==0.4.18
faiss-cpu==1.7.4

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.2
python-docx==1.1.0

# Web Interface
streamlit==1.29.0
fastapi==0.108.0
uvicorn==0.25.0

# Visualization
plotly==5.18.0
networkx==3.2.1
pyvis==0.3.2

# Utils
tqdm==4.66.1
pandas==2.1.4
numpy==1.26.2
